requirements.txt
asgiref==3.5.1
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==3.2.13
django-manifest-loader==1.0.0
djangorestframework==3.13.1
idna==3.3
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
requests==2.27.1
sqlparse==0.4.2
tinydb==4.7.0
urllib3==1.26.9
wrapt==1.14.1
